pyproject.toml
[tool.poetry]
name = "mqtt-pg-logger"
version = "0.1.0"
description = "MQTT PostgreSQL Logger"
authors = ["Tyler Matteson <tyler@agritheory.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.30.0"
attrs = "^24.2.0"
click = "^8.1.7"
jsonschema = "^4.23.0"
paho-mqtt = "^2.1.0"
pyyaml = "^6.0.2"
tzlocal = "^5.2"
aiomqtt = {git = "https://github.com/empicano/aiomqtt.git"}
quart = "^0.19.8"
quart-auth = "^0.10.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-postgresql = "^6.1.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99

[tool.isort]
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
indent = "\t"
